requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
pdfplumber==0.10.3
reportlab==4.0.9
pypdf==4.0.1
Werkzeug==3.0.1
gunicorn==21.2.0
python-dotenv==1.0.0


